+++ constants/InputConstant.ts
@@ -8,11 +8,17 @@ export const PLACEHOLDER = {
 
 export const ERROR_MESSAGE = {
   emailRequired: "이메일을 입력해 주세요.",
-  emailInvalid: "올바른 이메일 주소가 아닙니다.",
+  emailInvalid: "이메일 형식으로 작성해 주세요.",
   emailCheck: "이메일을 확인해 주세요.",
+
   passwordRequired: "비밀번호를 입력해 주세요.",
-  passwordInvalid: "영문, 숫자를 조합해 8자 이상 입력해 주세요.",
+  signinPasswordInvalid: "8자 이상 입력해 주세요.",
+  signupPasswordInvalid: "영문, 숫자를 조합해 8자 이상 입력해 주세요.",
   passwordCheck: "비밀번호를 확인해 주세요.",
+
+  confirmPasswordRequired: "비밀번호 확인을 입력해 주세요.",
+  confirmPasswordCheck: "비밀번호가 일치하지 않습니다.",
+
   nicknameRequired: "닉네임을 입력해 주세요.",
   nicknameInvalid: "10자 이하로 작성해 주세요.",
 };
@@ -22,10 +28,23 @@ export const EMAIL_RULES = {
   pattern: { value: /\S+@\S+\.\S+/, message: ERROR_MESSAGE.emailInvalid },
 };
 
-export const PASSWORD_RULES = {
+export const SIGNIN_PASSWORD_RULES = {
   required: ERROR_MESSAGE.passwordRequired,
   pattern: {
     value: /^(?=.*[A-Za-z])(?=.*\d).{8,}$/,
-    message: ERROR_MESSAGE.passwordInvalid,
+    message: ERROR_MESSAGE.signinPasswordInvalid,
   },
+};
+
+export const SIGNUP_PASSWORD_RULES = {
+  required: ERROR_MESSAGE.passwordRequired,
+  pattern: {
+    value: /^(?=.*[A-Za-z])(?=.*\d).{8,}$/,
+    message: ERROR_MESSAGE.signupPasswordInvalid,
+  },
+};
+
+export const NICKNAME_RULES = {
+  required: ERROR_MESSAGE.nicknameRequired,
+  maxLength: { value: 10, message: ERROR_MESSAGE.nicknameInvalid },
 };
