+++ pages/index.tsx
@@ -1,3 +1,6 @@
+import SideMenu from "@/components/SideMenu/SideMenu";
+import SignInForm from "@/components/SignForm/SignInForm";
+import SignUpForm from "@/components/SignForm/SignUpForm";
 import { Inter } from "next/font/google";
 import Head from "next/head";
 
@@ -12,6 +15,8 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
+      <SignInForm />
+      <SignUpForm />
     </>
   );
 }
